refactor: tea helper tool
