+++ app/tools/routes.py
@@ -2,8 +2,8 @@ from app import db
 from flask import jsonify, render_template, request
 from sqlalchemy import func
 
-from app.models import Character, Menu, liked_meals
-from app.tools.forms import DualCharacterForm
+from app.models import liked_meals, Character, Menu, TeaTopic
+from app.tools.forms import CharacterForm, DualCharacterForm
 from app.tools import bp
 from app.utility import get_choices
 
@@ -42,3 +42,40 @@ def get_meal_data():
     return jsonify({'meals': meals_data,
                     'meals_count': meals_count,
                     'num_chars': num_chars})
+
+@bp.route('/tea-helper', methods=['GET'])
+def tea_helper():
+    form = CharacterForm()
+    choices = get_choices(Character, True)
+    form.character.choices = choices
+
+    return render_template('tools/tea_helper.html', title='Tea Helper',
+                           page_name='Tea Party Helper', form=form)
+
+@bp.route('/get_tea_data', methods=['POST'])
+def get_tea_data():
+    cid = int(request.json['selected_option'])
+    char = db.session.get(Character, cid)
+
+    query = char.likes_tea.select()
+    tea = db.session.scalars(query).all()
+    tea_data = [{'tea': t.name} for t in tea]
+    tea_count = len(tea)
+
+    query = char.likes_tea_topic.select().order_by(TeaTopic.data)
+    topics = db.session.scalars(query).all()
+    topic_data = [{'topic': t.data} for t in topics]
+
+    final_topics = sorted(char.final_tea_comment, key=lambda final_topic: final_topic.comment)
+    comment_data = []
+    answer_data = []
+    for topic in final_topics:
+        comment_data.append({'comment': topic.comment})
+        answer_data.append({'answer': topic.response})
+
+    return jsonify({'tea': tea_data,
+                    'tea_count': tea_count,
+                    'topics': topic_data,
+                    'comments': comment_data,
+                    'answers': answer_data})
+
